fix session circular dependince
