+++ src/app/api/feedback/feedback.routes.ts
@@ -1,9 +1,7 @@
 import { Router } from '@lib/methods';
-import { CrudService } from '@shared/crud';
 import { FeedbackSchema, FeedbackModel } from './feedback.model';
-import { Repo } from '@shared/crud/crud.repo';
+import { Repo, CrudRouter, CrudService } from '@shared/crud';
 import { Constants } from '@core/helpers';
-import { CrudRouter } from '@shared/crud/crud.router';
 
 @Router(Constants.Endpoints.Feedback)
 export class FeedbackRouter extends CrudRouter<FeedbackSchema> {
