+++ src/app/api/sessions/sessions.routes.ts
@@ -1,13 +1,13 @@
 import { SessionSchema } from './sessions.model';
 import { Response, Request } from 'express';
-import { Auth } from '@api/portal';
+import { Auth } from '../portal/auth';
 import { Get, Router, Patch } from '@lib/methods';
 import { tokenService, SuccessResponse, Constants } from '@core/helpers';
-import { CrudRouter } from '@shared/crud';
-import { sessionsService, SessionService } from './sessions.service';
+import { CrudRouter } from '../../shared/crud';
+import { sessionsService, SessionsService } from './sessions.service';
 
 @Router(Constants.Endpoints.SESSIONS)
-export class SessionRouter extends CrudRouter<SessionSchema, SessionService> {
+export class SessionRouter extends CrudRouter<SessionSchema, SessionsService> {
     constructor() {
         super(sessionsService);
     }
